fix: checkGit() called for non-git repos
